primeira versao ordem de serviço service

+++ Gerenciador_Manutencao/Model/DTO/RequestDTO/ItemRequestDTO.cs
@@ -3,5 +3,5 @@ public class ItemRequestDTO
     public string Tipo { get; set; }
     public int UnidadeDeMedida { get; set; }
     public string Descricao { get; set; }
-    public int Quantidade { get; set; }
+    public float Quantidade { get; set; }
 }
